refactor(mpcTask): 修改上传任务时各方partInfo字段

+++ DAVEX_base/src/main/java/DavexBase/entity/MpcTask.java
@@ -37,7 +37,7 @@ public class MpcTask {
   @TableField(typeHandler = FastjsonTypeHandler.class)
   private JSONObject runtimeParameters;
   private Integer N;
-  private Integer part;
+  private Long part;
   private String host;
   private Integer port;
   @JsonInclude(Include.NON_NULL)
